requirements.txt
django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0